backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
email-validator
python-multipart
cryptography
# Admin Panel & UI
starlette-admin
sqlalchemy-utils
itsdangerous